python/requirements.txt
openai
flask
flask-cors
requests
pyyaml
